Push Connect-specific hooks from IntegrationGateway to IntegrationConnector
IntegrationGateway accumulated eight Connect-mode hooks over time, all
returning types that already lived under connector/. Now that the
capability model has IntegrationConnector and IntegrationImporter as
peer capability classes (cf. the recent IntegrationManageViewPane
cleanup), these hooks belong on the connector — not on the
capability-neutral gateway. Import-only integrations should not be
forced to stub them out.

Methods moved from IntegrationGateway to IntegrationConnector:
- get_monitor()
- get_controller()
- get_health_status_provider()
- get_entity_video_stream()
- get_entity_video_snapshot()
- get_sensor_response_video_stream()
- get_sensor_response_event_snapshot_url()
- get_external_view_data()

Per-service implementations relocated from each *Gateway (integration.py)
to the corresponding *Connector.

Production call sites updated to chain through get_connector() with
None-guards where the caller might encounter an Import-only or
connector-less integration. The three Connect-side view sites trust the
structural invariant. The do_control path raises a clear RuntimeError
if no connector or controller is available.

File renames to match the relocated class names:
- frigate_sync.py -> frigate_connector.py (FrigateConnector)
- hass_sync.py -> hass_connector.py (HassConnector)
- zm_sync.py -> zm_connector.py (ZmConnector)
- corresponding test files

IntegrationGateway's surface is now its capability-neutral core:
get_metadata, notify_settings_changed, validate_configuration,
validate_access, get_connector, get_importer.
