+++ src/hi/apps/control/tests/test_one_click_control_service.py
@@ -217,7 +217,7 @@ class TestOneClickFindController(BaseTestCase):
         mock_integration_controller = Mock()
         mock_integration_controller.do_control.return_value = mock_control_result
         mock_integration_gateway = Mock()
-        mock_integration_gateway.get_controller.return_value = mock_integration_controller
+        mock_integration_gateway.get_connector.return_value.get_controller.return_value = mock_integration_controller
         mock_integration_manager = Mock()
         mock_integration_manager.get_integration_gateway.return_value = mock_integration_gateway
 
@@ -264,7 +264,7 @@ class TestOneClickFindController(BaseTestCase):
         mock_integration_controller = Mock()
         mock_integration_controller.do_control.return_value = mock_control_result
         mock_integration_gateway = Mock()
-        mock_integration_gateway.get_controller.return_value = mock_integration_controller
+        mock_integration_gateway.get_connector.return_value.get_controller.return_value = mock_integration_controller
         mock_integration_manager = Mock()
         mock_integration_manager.get_integration_gateway.return_value = mock_integration_gateway
