+++ src/hi/services/hass/integration.py
@@ -1,12 +1,8 @@
 import logging
 from typing import List, Optional
 
-from hi.apps.entity.models import Entity
-from hi.apps.entity.transient_models import VideoSnapshot
 from hi.apps.system.enums import HealthStatusType
-from hi.apps.system.health_status_provider import HealthStatusProvider
 
-from hi.integrations.connector.integration_controller import IntegrationController
 from hi.integrations.integration_gateway import IntegrationGateway
 from hi.integrations.connector.integration_connector import IntegrationConnector
 from hi.integrations.models import IntegrationAttribute
@@ -15,13 +11,10 @@ from hi.integrations.transient_models import (
     IntegrationMetaData,
     IntegrationValidationResult,
 )
-from hi.apps.monitor.periodic_monitor import PeriodicMonitor
 
-from .hass_controller import HassController
 from .hass_manager import HassManager
 from .hass_metadata import HassMetaData
-from .hass_sync import HassConnector
-from .monitors import HassMonitor
+from .hass_connector import HassConnector
 
 logger = logging.getLogger(__name__)
 
@@ -31,15 +24,9 @@ class HassGateway( IntegrationGateway ):
     def get_metadata(self) -> IntegrationMetaData:
         return HassMetaData
 
-    def get_monitor(self) -> PeriodicMonitor:
-        return HassMonitor()
-    
-    def get_controller(self) -> IntegrationController:
-        return HassController()
-    
     def notify_settings_changed(self):
         """Notify HASS integration that settings have changed.
-        
+
         Delegates to HassManager to reload configuration and notify monitors.
         """
         try:
@@ -48,9 +35,6 @@ class HassGateway( IntegrationGateway ):
             logger.debug('HASS integration notified of settings change')
         except Exception as e:
             logger.exception(f'Error notifying HASS integration of settings change: {e}')
-    
-    def get_health_status_provider(self) -> HealthStatusProvider:
-        return HassManager()
 
     def get_connector(self) -> IntegrationConnector:
         return HassConnector()
@@ -84,38 +68,3 @@ class HassGateway( IntegrationGateway ):
         except Exception as e:
             logger.exception(f'Error in HASS access validation: {e}')
             return ConnectionTestResult.failure(f'Access validation error: {e}')
-
-    def get_entity_video_snapshot(self, entity: Entity) -> Optional[VideoSnapshot]:
-        if not entity.has_video_snapshot:
-            return None
-        if entity.integration_id != HassMetaData.integration_id:
-            return None
-
-        hass_manager = HassManager()
-        # ``Entity.integration_name`` is the HassDevice device_id (an HI
-        # grouping construct), not the HA state id the attrs cache is
-        # keyed by. The manager bridges the two via a sync-time-built
-        # map of HI Entity.id -> camera-domain HA state id.
-        ha_state_id = hass_manager.get_ha_state_id_for_entity( entity )
-        if not ha_state_id:
-            return None
-
-        attrs = hass_manager.get_latest_attrs( ha_state_id )
-        if not attrs:
-            return None
-
-        entity_picture = attrs.get( 'entity_picture' )
-        if not entity_picture:
-            return None
-
-        # Some HA integrations emit an absolute URL; pass those
-        # through unchanged. Relative paths get the HA base prefix.
-        if entity_picture.startswith( ('http://', 'https://') ):
-            source_url = entity_picture
-        else:
-            client = hass_manager.hass_client
-            if not client:
-                return None
-            source_url = f'{client.api_base_url}{entity_picture}'
-
-        return VideoSnapshot( source_url = source_url )
