+++ src/hi/services/zoneminder/tests/test_zm_connector.py
@@ -14,7 +14,7 @@ from hi.integrations.integration_manager import IntegrationManager
 from hi.integrations.connector.sync_result import IntegrationSyncResult
 from hi.integrations.transient_models import IntegrationKey
 
-from hi.services.zoneminder.zm_sync import ZmConnector
+from hi.services.zoneminder.zm_connector import ZmConnector
 from hi.services.zoneminder.zm_metadata import ZmMetaData
 
 logging.disable(logging.CRITICAL)
@@ -141,8 +141,8 @@ class TestZmConnectorStateSync(TestCase):
         )
         self.synchronizer._zm_manager = self.mock_manager
     
-    @patch('hi.services.zoneminder.zm_sync.Entity.objects.filter_by_integration_key')
-    @patch('hi.services.zoneminder.zm_sync.Sensor.objects.filter_by_integration_key')
+    @patch('hi.services.zoneminder.zm_connector.Entity.objects.filter_by_integration_key')
+    @patch('hi.services.zoneminder.zm_connector.Sensor.objects.filter_by_integration_key')
     @patch.object(ZmConnector, '_create_zm_entity')
     def test_sync_states_creates_zm_entity_when_missing(self, mock_create_entity, mock_sensor_filter, mock_entity_filter):
         """Test _sync_states creates ZM entity when it doesn't exist and verifies entity creation parameters"""
@@ -189,8 +189,8 @@ class TestZmConnectorStateSync(TestCase):
         # Test behavior: should not have errors when successful
         self.assertEqual(len(result.error_list), 0)
     
-    @patch('hi.services.zoneminder.zm_sync.Entity.objects.filter_by_integration_key')
-    @patch('hi.services.zoneminder.zm_sync.Sensor.objects.filter_by_integration_key')
+    @patch('hi.services.zoneminder.zm_connector.Entity.objects.filter_by_integration_key')
+    @patch('hi.services.zoneminder.zm_connector.Sensor.objects.filter_by_integration_key')
     def test_sync_states_missing_sensor_error(self, mock_sensor_filter, mock_entity_filter):
         """Test _sync_states handles missing run state sensor and stops processing gracefully"""
         # Mock states
@@ -217,8 +217,8 @@ class TestZmConnectorStateSync(TestCase):
         # Should not have created any entities since sensor missing
         self.assertEqual(len(result.info_list), 0)
     
-    @patch('hi.services.zoneminder.zm_sync.Entity.objects.filter_by_integration_key')
-    @patch('hi.services.zoneminder.zm_sync.Sensor.objects.filter_by_integration_key')
+    @patch('hi.services.zoneminder.zm_connector.Entity.objects.filter_by_integration_key')
+    @patch('hi.services.zoneminder.zm_connector.Sensor.objects.filter_by_integration_key')
     def test_sync_states_updates_value_range_when_changed(self, mock_sensor_filter, mock_entity_filter):
         """Test _sync_states updates value range when states change and persists changes"""
         # Mock new states
@@ -264,8 +264,8 @@ class TestZmConnectorStateSync(TestCase):
         self.assertIn('start', message)
         self.assertIn('pause', message)
     
-    @patch('hi.services.zoneminder.zm_sync.Entity.objects.filter_by_integration_key')
-    @patch('hi.services.zoneminder.zm_sync.Sensor.objects.filter_by_integration_key')
+    @patch('hi.services.zoneminder.zm_connector.Entity.objects.filter_by_integration_key')
+    @patch('hi.services.zoneminder.zm_connector.Sensor.objects.filter_by_integration_key')
     def test_sync_states_no_update_when_unchanged(self, mock_sensor_filter, mock_entity_filter):
         """Test _sync_states doesn't update when state values unchanged and preserves existing state"""
         # Mock states - identical to existing
@@ -488,7 +488,7 @@ class TestZmConnectorExistingEntities(TestCase):
         self.mock_manager.ZM_MONITOR_INTEGRATION_NAME_PREFIX = 'monitor'
         self.synchronizer._zm_manager = self.mock_manager
     
-    @patch('hi.services.zoneminder.zm_sync.Entity.objects.filter')
+    @patch('hi.services.zoneminder.zm_connector.Entity.objects.filter')
     def test_get_existing_zm_monitor_entities_filters_by_integration_id(self, mock_filter):
         """Test _get_existing_zm_monitor_entities filters by correct integration ID"""
         mock_filter.return_value = []
@@ -498,7 +498,7 @@ class TestZmConnectorExistingEntities(TestCase):
         
         mock_filter.assert_called_once_with(integration_id=ZmMetaData.integration_id)
     
-    @patch('hi.services.zoneminder.zm_sync.Entity.objects.filter')
+    @patch('hi.services.zoneminder.zm_connector.Entity.objects.filter')
     def test_get_existing_zm_monitor_entities_handles_missing_integration_key(self, mock_filter):
         """Test entity retrieval handles entities without integration keys"""
         # Mock entity without integration key
@@ -516,7 +516,7 @@ class TestZmConnectorExistingEntities(TestCase):
         # Should NOT include entity in result (mock key doesn't start with 'monitor' prefix)
         self.assertEqual(len(result_dict), 0)
     
-    @patch('hi.services.zoneminder.zm_sync.Entity.objects.filter')
+    @patch('hi.services.zoneminder.zm_connector.Entity.objects.filter')
     def test_get_existing_zm_monitor_entities_filters_monitor_entities(self, mock_filter):
         """Test entity retrieval only includes monitor entities"""
         # Mock entities - one monitor, one non-monitor
@@ -703,8 +703,8 @@ class TestZmConnectorWithRealData(TestCase):
             mock_monitors.append(mock_monitor)
         return mock_monitors
     
-    @patch('hi.services.zoneminder.zm_sync.Entity.objects.filter_by_integration_key')
-    @patch('hi.services.zoneminder.zm_sync.Sensor.objects.filter_by_integration_key')
+    @patch('hi.services.zoneminder.zm_connector.Entity.objects.filter_by_integration_key')
+    @patch('hi.services.zoneminder.zm_connector.Sensor.objects.filter_by_integration_key')
     def test_sync_states_with_real_zm_state_names(self, mock_sensor_filter, mock_entity_filter):
         """Test state sync handles real ZM state names: default, Away, HomeDay, Disabled"""
         # Use real state data
@@ -814,7 +814,7 @@ class TestZmConnectorWithRealData(TestCase):
         mock_entity.save.assert_not_called()
         self.assertEqual(result.updated_list, [])
     
-    @patch('hi.services.zoneminder.zm_sync.Entity.objects.filter')
+    @patch('hi.services.zoneminder.zm_connector.Entity.objects.filter')
     def test_get_existing_entities_with_real_monitor_id_patterns(self, mock_filter):
         """Test existing entity retrieval with realistic monitor ID patterns"""
         # Create mock entities with integration keys matching real monitor IDs
@@ -875,10 +875,10 @@ class TestZmConnectorWithRealData(TestCase):
             # Verify IsActive is boolean-like string
             self.assertIn(state['IsActive'], ['0', '1'])
     
-    @patch('hi.services.zoneminder.zm_sync.HiModelHelper.create_movement_sensor')
-    @patch('hi.services.zoneminder.zm_sync.HiModelHelper.create_discrete_controller')
-    @patch('hi.services.zoneminder.zm_sync.HiModelHelper.create_movement_event_definition')
-    @patch('hi.services.zoneminder.zm_sync.transaction.atomic')
+    @patch('hi.services.zoneminder.zm_connector.HiModelHelper.create_movement_sensor')
+    @patch('hi.services.zoneminder.zm_connector.HiModelHelper.create_discrete_controller')
+    @patch('hi.services.zoneminder.zm_connector.HiModelHelper.create_movement_event_definition')
+    @patch('hi.services.zoneminder.zm_connector.transaction.atomic')
     @patch.object(Entity, 'save')
     def test_create_monitor_entity_with_real_monitor_variations(self, mock_save, mock_atomic,
                                                                 mock_create_event, mock_create_controller,
